requirements.txt
requests==2.27.1
pandas==1.3.5
geopandas==0.9.0
matplotlib-base==3.5.0
owslib==0.24.1
